seg fault fix

+++ driver.cpp
@@ -8,7 +8,7 @@ int main(){
  string name;
 cout << "Enter your name below. No spaces." << endl;
 cin>>name;
-float score=bot1.get_score(name);
+int score=bot1.get_score(name);
 
 cout << bot1.get_history(score) << endl;
